requirement.txt
alembic==1.1.0.dev0
apturl==0.5.2
asgiref==3.3.1
astroid==2.4.2
Babel==2.6.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.0rc1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.4
django-cors-headers==3.6.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
Flask==1.1.1
Flask-BabelEx==0.9.3
Flask-Compress==1.4.0
Flask-Gravatar==0.4.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security-Too==3.4.2
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
future==0.18.2
html5lib==1.0.1
httplib2==0.14.0
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.6.4
itsdangerous==1.1.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
kazam==1.4.5
keyring==18.0.1
keyrings.alt==3.4.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
ldap3==2.4.1
libvirt-python==6.1.0
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
netifaces==0.10.4
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
parso==0.7.1
passlib==1.7.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
protobuf==3.6.1
psutil==5.5.1
psycopg2==2.8.6
pyasn1==0.4.2
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.7.3
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
PyQt5==5.14.1
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.2
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2020.4
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.25.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
scour==0.37
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==1.9.5
speaklater==1.4
SQLAlchemy==1.3.12
sqlparse==0.4.1
sshtunnel==0.1.4
systemd-python==234
toml==0.10.2
traitlets==5.0.5
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.2
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.1
wrapt==1.12.1
WTForms==2.2.1
xkit==0.0.0
